Add ApproverSettings endpoint and enhance related codebase
Enhanced the codebase by adding ApproverSettings endpoints and interlacing these changes into the existing routes and controllers. This update was necessary to accommodate recent user role configurations and other adjustments to the application's access and authorization matrix. Removed deprecated code from the User model, added attribute transformation in the LocationController, and updated the role permissions within the AuthController. Also improved the validation in the DepartmentController to account for lack of locations. Furthermore, upon archiving a user, the system now checks for any existing user-approver relationships that might still be active. Commented-out parts of old code are preserved for reference purposes. Finally, added 'unique_in_array' custom validation rule to AppServiceProvider.

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use App\Models\Sedar;
 use App\Models\Module;
 use App\Models\Department;
+use App\Models\UserApprover;
 use Illuminate\Http\Request;
 use App\Models\Access_Permission;
 use App\Http\Requests\UserRequest;
@@ -226,36 +227,41 @@ class UserController extends Controller
     }
 
     public function archived(UserRequest $request, $id)
-    {
-        $auth_id = auth('sanctum')->user()->id;
-        if ($id == $auth_id) {
-            return response()->json(['message' => 'Unable to Archive, User already in used!'], 409);
-        }
-        $status = $request->status;
-        $User = User::query();
-        if (!$User->withTrashed()->where('id', $id)->exists()) {
-            return response()->json(['error' => 'User Route Not Found'], 404);
-        }
-        if ($status == false) {
-            if (!User::where('id', $id)->where('is_active', true)->exists()) {
-                return response()->json(['message' => 'No Changes'], 200);
-            } else {
-                $updateStatus = $User->where('id', $id)->update(['is_active' => false]);
-                $User->where('id', $id)->delete();
-                return response()->json(['message' => 'Successfully Deactived!'], 200);
-            }
-        }
-        if ($status == true) {
-            if (User::where('id', $id)->where('is_active', true)->exists()) {
-                return response()->json(['message' => 'No Changes'], 200);
-            } else {
-                $restoreUser = $User->withTrashed()->where('id', $id)->restore();
-                $updateStatus = $User->update(['is_active' => true]);
-                return response()->json(['message' => 'Successfully Activated!'], 200);
-            }
-        }
-
+{
+    $auth_id = auth('sanctum')->user()->id;
+    if ($id == $auth_id) {
+        return response()->json(['message' => 'Unable to Archive, User already in used!'], 409);
     }
+    $status = $request->status;
+    $User = User::query();
+    if (!$User->withTrashed()->where('id', $id)->exists()) {
+        return response()->json(['error' => 'User Route Not Found'], 404);
+    }
+    if ($status == false) {
+        if (!User::where('id', $id)->where('is_active', true)->exists()) {
+            return response()->json(['message' => 'No Changes'], 200);
+        } else {
+            $userSettingApproverCheck = UserApprover::where('requester_id', $id)->orWhere('approver_id', $id)->exists();
+            if ($userSettingApproverCheck) {
+                return response()->json(['message' => 'User Account still in use'], 409);
+            }
+
+            $updateStatus = $User->where('id', $id)->update(['is_active' => false]);
+            $User->where('id', $id)->delete();
+            return response()->json(['message' => 'Successfully Deactived!'], 200);
+        }
+    }
+    if ($status == true) {
+        if (User::where('id', $id)->where('is_active', true)->exists()) {
+            return response()->json(['message' => 'No Changes'], 200);
+        } else {
+            $restoreUser = $User->withTrashed()->where('id', $id)->restore();
+            $updateStatus = $User->update(['is_active' => true]);
+            return response()->json(['message' => 'Successfully Activated!'], 200);
+        }
+    }
+
+}
 
     public function test(Request $request)
     {
